changed navbar header + file names
file names changed to be more consistent
