+++ app/src/App.css
@@ -18,7 +18,7 @@
   position: fixed;
   top: 0;
   left: 0;
-  height: 100px;
+  height: 80px;
   width: 100%;
   background-color: #222;
   padding: 0;
@@ -30,31 +30,29 @@
   align-items: center;
 }
 
+.dropdown {
+  position: relative;
+  display: inline-block;
+  font-family: sans-serif;
+  height: 100%;
+  margin-left:8%;
 
-.header-buttons{
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-around;
-    width: 100%;
-    color:white;
 }
 
+.dropdown-button {
+  position: relative;
+  text-decoration: none;
+  font-size: 25px;
+  background-color:#222;
 
-.header-button{
-position: relative;
-text-decoration: none;
-padding: 5px 5px 5px 5px;
-font-size: 25px;
+  border: none;
+  height: 100%;
+  padding-left: 40px;
+  padding-right: 40px;
+
+  color:white;
 }
-
-.header-button:hover {
-color: rgb(0, 165, 206);
-cursor:pointer;
-transition:color .3s;
-}
-
-.header-button::before {
+.dropdown-button:before {
 content: "";
 position: absolute;
 display: block;
@@ -67,6 +65,75 @@ transform: scaleX(0);
 transition: transform 0.3s ease;
 }
 
-.header-button:hover::before {
+.dropdown-button:hover{
+  transform: scaleX(1);
+}
+
+.dropdown-content {
+  opacity:0;
+  position: absolute;
+  background-color: #222;
+  min-width: 160px;
+  box-shadow: 0 8px 16px rgba(0,0,0,0.2);
+  z-index: 1;
+  border-radius: 4px;
+  font-size: 20px;
+  transition: opacity .2s ease-out;
+}
+
+.dropdown-content div{
+  color: rgba(255, 255, 255, 0.8);
+  padding: 10px 14px;
+  display: block;
+  
+}
+
+.dropdown-content div:hover {
+  background-color: #1d1d1d;
+}
+
+/* Show dropdown on hover */
+.dropdown:hover .dropdown-content {
+  opacity: 1;
+}
+
+
+
+.dropdown-options{
+position: relative;
+text-decoration: none;
+padding: 5px 5px 5px 5px;
+font-size: 25px;
+}
+
+.dropdown-options:hover {
+color: rgb(0, 165, 206);
+cursor:pointer;
+transition:color .3s;
+}
+
+.dropdown-options::before {
+content: "";
+position: absolute;
+display: block;
+width: 100%;
+height: 1.5px;
+bottom: 0;
+left: 0;
+background-color: rgb(0, 165, 206);
+transform: scaleX(0);
+transition: transform 0.3s ease;
+}
+
+.dropwdown-options:hover::before {
 transform: scaleX(1);
+}
+
+.page-title{
+  color:white;
+  position: fixed;
+  left: 50%;
+  transform: translate(-50%, 0);
+  font-size: 25px;
+
 }
