+++ app/src/components/Navbar.js
@@ -1,37 +1,48 @@
-import { Link } from 'react-router-dom';
 import { useNavigate } from 'react-router-dom';
+import React, {useState} from 'react';
 
 const Navbar = () => {
 
+  const [page, setPage] = useState("Home");
 
   const navigate = useNavigate();
 
   const homenav = () => {
     navigate("/home");
+    setPage('Home');
   };
   const bennav = () => {
     navigate("/ben");
+    setPage('Ben\'s Page');
   };
   const willnav = () => {
     navigate("/will");
+    setPage('Will\'s Page');
   };
     const imagenav = () => {
     navigate("/images");
+    setPage('Images');
   };
 
     const ben_apinav = () => {
     navigate("/BenAPI");
+    setPage('Ben\'s API');
   };
   return (
     <div className='navBar'>
 
-                <div class="header-buttons">
-                <div class="header-button" onClick={homenav}> Home</div>
-                <div class="header-button" onClick={bennav}>Ben's Page</div>
-                <div class="header-button" onClick={willnav}>Will's Page</div>
-                <div class="header-button"onClick={imagenav}>Images</div>
-                <div class="header-button"onClick={ben_apinav}>Ben's API</div>
-                </div> 
+      <div className="dropdown">
+      <button className="dropdown-button">Pages ▾</button>
+      <div className="dropdown-content">
+        <div onClick={homenav} className="dropdown-options">Home</div>
+        <div  onClick={bennav} className="dropdown-options">Ben's Page</div>
+        <div  onClick={willnav} className="dropdown-options">Will's Page</div>
+        <div onClick={imagenav} className="dropdown-options">Images</div>
+        <div onClick={ben_apinav} className="dropdown-options">Ben's API</div>
+      </div>
+    </div>
+
+    <div className="page-title">{page}</div>
     </div>
   );
 };
