+++ app/src/pages/BenAPI.js
@@ -8,6 +8,7 @@ function BenAPI(){
   const [colors, setColors] = useState([]);
 
   useEffect(() => { fetchColor();}, []);
+
   const fetchColor = async() => {
       const res = await fetch('https://api.sampleapis.com/csscolornames/colors')
       const data = await res.json();
@@ -22,6 +23,7 @@ function BenAPI(){
     },
     body: JSON.stringify({ name: colorName, hex: hex })
   });
+  
   if (res.ok){
     setHex('');
     setColorName('');
@@ -34,42 +36,42 @@ function BenAPI(){
     <div className="main">
       <h1 className="header">CSS Colors</h1>
 
-       <div>
-      <input
-        type="text"
-        value={colorName}
-        onChange={(e) => setColorName(e.target.value)}
-        placeholder='Black'
-      />
-      <input
-        type="text"
-        value={hex}
-        onChange={(e) => setHex(e.target.value)}
-        placeholder='#000000'
-      />
-      <button onClick={handlePost}>Add</button>
+      <div>
+        <input
+          type="text"
+          value={colorName}
+          onChange={(e) => setColorName(e.target.value)}
+          placeholder='Black'
+        />
+        <input
+          type="text"
+          value={hex}
+          onChange={(e) => setHex(e.target.value)}
+          placeholder='#000000'
+        />
+        <button onClick={handlePost}>Add</button>
+      </div>
+
+
+      <div className="container">
+        {colors.map((color, index) => (
+          <div
+            key={index}
+            style={{
+              backgroundColor: color.hex,
+              color: '#fff',
+              padding: '5px',
+              borderRadius: '8px',
+              width: '80px',
+              textAlign: 'center',
+              fontSize: '10px'
+            }}
+          >
+            {color.name}
+          </div>
+        ))}
+      </div>
     </div>
-
-
-        <div className="container">
-      {colors.map((color, index) => (
-        <div
-          key={index}
-          style={{
-            backgroundColor: color.hex,
-            color: '#fff',
-            padding: '5px',
-            borderRadius: '8px',
-            width: '80px',
-            textAlign: 'center',
-            fontSize: '10px'
-          }}
-        >
-          {color.name}
-        </div>
-      ))}
-        </div>
-        </div>
   );
 }
